`nix flake check` expects overlays to be an attrset.
This is an odd choice by upstream as it means that anyone who wants to
use the overlays from this package will need to convert it to a list
first, but oh well.

Anyway, we move some useful lib code out of the nix/overlays directory
and into flake.nix, where we use it to build an `overlays` attrset,
then merge that lib code back into the overlays attrset manually, so
that downstream users can take advantage of it.