Export a `lib` attribute.
This is non-standard, but super helpful when you just want to make use
of some of these library functions in a system-independent way.

Note that the `lib` attribute evaluates nixpkgs with our local overlay
at the `x86_64-linux` system attribute. This appears to work fine on
non-x86_64-linux systems, because I don't think there's anything in
lib that is system-dependent.