nixosConfigurations is a special attribute.
It's not properly documented, but according to the current
implementation here:

https://github.com/NixOS/nix/blob/13897afbe6cf7ef8013c0c94109696bb7b13d0c0/src/nix/flake.cc

the `nixosConfigurations` attribute is not meant to be per-system as,
e.g., the `packages` attribute is meant to be. It's likely that
Flake-aware applications that operate on the `nixosConfigurations`
attribute, such as deploy-rs, depend on this structure, so this commit
flattens the `nixosConfigurations` attribute.

(We do the same for `darwinConfigurations` which, while it's not an
official Flake attribute, is special-cased by nix-darwin.)

`hydraJobs` don't need to be per-system, either, but they can be, and
it's helpful to do so for other CI systems such as Hercules CI.
Therefore, we add these attributes to `self.hydraJobs` as
`self.hydraJobs.nixosConfigurations.x86_64-linux` and
`self.hydraJobs.darwinConfigurations.x86_64-darwin`.

(In the future, we may need to further partition these `hydraJobs`
attributes by their actual system types. At the moment, all of the
NixOS systems we build are x86_64-linux, and all of the macOS systems
we build are x86_64-darwin, so this simplified implementation is fine
for now.)